+++ jvm/src/main/scala/com/nawforce/apexlink/api/Org.scala
@@ -314,6 +314,7 @@ object Org {
     options.indexerConfiguration.foreach(values =>
       ServerOps.setIndexerConfiguration(IndexerConfiguration(values._1, values._2))
     )
+    val blockPrefetchThreads = ServerOps.resolveBlockPrefetchThreads(options.blockPrefetchThreads)
     options.autoFlush.foreach(enabled => ServerOps.setAutoFlush(enabled))
     options.cache.foreach(enabled =>
       if (!enabled) {
@@ -332,7 +333,8 @@ object Org {
         issueManager,
         workspace,
         unusedEnabled = options.unused.getOrElse(true),
-        unusedOnError = options.unusedOnError.getOrElse(false)
+        unusedOnError = options.unusedOnError.getOrElse(false),
+        blockPrefetchThreads = blockPrefetchThreads
       )
     }
   }